chore: move authentication state and server state to the api client (wip)
