+++ packages/api-client/package.json
@@ -64,6 +64,7 @@
     "@types/content-type": "^1.1.6",
     "@vitejs/plugin-vue": "^4.4.0",
     "@vitest/coverage-v8": "^0.34.4",
+    "openapi-types": "^12.1.3",
     "tsc-alias": "^1.8.8",
     "vite": "^4.5.2",
     "vite-plugin-css-injected-by-js": "^3.3.0",
