+++ packages/api-client/src/index.ts
@@ -2,7 +2,6 @@ export { ApiClient } from './components/ApiClient'
 export { default as HttpMethod } from './components/HttpMethod.vue'
 
 export * from './helpers'
-export * from './stores/apiClientStore'
-export * from './stores/requestStore'
+export * from './stores'
 export * from './types'
 export * from './fixtures'
